requirements.txt
# requirements.txt
pandas
python-docx
langchain-chroma
langchain
langgraph
langchain_community
google-generativeai
openai
sentence-transformers
pydantic
python-dotenv
PyYAML
tqdm
loguru
protobuf<3.21
grpcio-status==1.48.2
torch
transformers
rank_bm25
markitdown[docx]
underthesea
pyvi
langchain-huggingface
streamlit
redis
aioredis
langgraph-checkpoint-redis
langmem
dotenv